Reformat to properly generate data; add logging
The pipeline is now able to run for long periods of time. Fixed an error
where all frames have the same id (that of the tfrecord). Added an overwrite
option for writing data, because it writes to the same folder as previous
data. Added logging for easy debug. Should pipe to a file later, since this
is running for long periods. Added a 25-pt threshold for clusters. Probably
did some other stuff that I forgot. Not adding data to git yet . . . let's
discuss that.